Fix for scale > 46000 input generation. Thanks to @dii14. Closes #39

+++ include/puzzler/puzzles/mining.hpp
@@ -146,7 +146,7 @@ namespace puzzler
       if(scale <=4){
         params->threshold=0xFFFFFFFFFFFFFFFFull;
       }else{
-        params->threshold=(uint64_t)(pow(2.0,64) / (scale*scale));
+        params->threshold=(uint64_t)(pow(2.0,64) / (scale*(double)scale));
       }
 
       return params;
